fix test

+++ tests/go-e2e-statfs/main.go
@@ -26,7 +26,7 @@ func main() {
 		"files":   statfs.Files,
 		"flags":   statfs.Flags,
 		"frsize":  statfs.Frsize,
-		"fsid":    []int64{int64(statfs.Fsid.X__val[0]), int64(statfs.Fsid.X__val[1])}, // Convert fsid to list
+		"fsid":    []int32{int32(statfs.Fsid.X__val[0]), int32(statfs.Fsid.X__val[1])}, // Convert fsid to list
 		"namelen": statfs.Namelen,
 		"spare":   statfs.Spare,
 		"type":    statfs.Type,
